Illustrate why mounting cacheFolder doesn't work
When mounting globalFolder, the `yarn install` step in the Containerfile
will copy everything from the global cache to the project-local cache.
That means the project-local cache stays populated after the build.

When mounting cacheFolder, `yarn install` doesn't do anything (the cache
is already populated). But when the build ends, podman unmounts the
folder and the cache disappears. When you try to run the app:

    Error: Required package missing from disk.
    If you keep your packages inside your repository then restarting the Node process may be enough.
    Otherwise, try to run an install first.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>